Build bcp if necessary

+++ engine/external_libraries/copy_boost.sh
@@ -13,12 +13,23 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-boost_src="$1"
 src_dirs="include platform src tests"
 out_dir="./external_libraries/boost"
 
+if [ -z "$1" ]; then
+    echo "Usage: `basename $0` BOOST_SOURCE_DIR"
+    exit 1
+fi
+
+boost_src="$1"
+bcp="${boost_src}/dist/bin/bcp"
+
+if [ ! -x "$bcp" ]; then
+    ( cd "$boost_src" && ./bootstrap.sh && ./b2 tools/bcp ) || exit 1
+fi
+
 mkdir -p "$out_dir"
-"$boost_src/dist/bin/bcp" --boost="$boost_src" --scan `find $src_dirs -name '*.hpp' -o -name '*.cpp'` "$out_dir"
+"$bcp" --boost="$boost_src" --scan `find $src_dirs -name '*.hpp' -o -name '*.cpp'` "$out_dir"
 
 rm -f "$out_dir/Jamroot"
